llvm: Move code generation to a separate file
